User api and authentication set up

+++ backend/controllers/recipesController.js
@@ -3,12 +3,14 @@ const { globalAgent } = require('http')
 const { create } = require('../models/recipeModel')
 
 const Recipe = require('../models/recipeModel')
+const User = require('../models/userModel')
+
 
 // @desc Get recipes
 // @route GET /api/recipes
 // @access Private
 const getRecipes = asyncHandler(async (req, res) => {
-    const recipes = await Recipe.find()
+    const recipes = await Recipe.find({ user: req.user.id })
     res.status(200).json(recipes)
 })
 
@@ -23,6 +25,7 @@ const createRecipes = asyncHandler(async (req, res) => {
     }
 
     const recipe = await Recipe.create({
+        user: req.user.id,
         name: req.body.name,
         timeDuration: req.body.timeDuration,
         category: req.body.category,
@@ -31,6 +34,7 @@ const createRecipes = asyncHandler(async (req, res) => {
         nutrition: req.body.nutrition,
         numOfServings: req.body.numOfServings
     })
+    // console.log(req.user.id)
     res.status(200).json(recipe)
 })
 
@@ -38,10 +42,23 @@ const createRecipes = asyncHandler(async (req, res) => {
 // @route PUT /api/recipes
 // @access Private
 const updateRecipes = asyncHandler(async (req, res) => {
-    const recipe = Recipe.findById(req.params.id)
+    const recipe = await Recipe.findById(req.params.id)
     if(!recipe) {
         res.status(400)
         throw new Error('Recipe not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the
+    // recipe user
+    if(recipe.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedRecipe = await Recipe.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -53,11 +70,26 @@ const updateRecipes = asyncHandler(async (req, res) => {
 // @route DELETE /api/recipes/:id
 // @access Private
 const deleteRecipes = asyncHandler(async (req, res) => {
-    const recipe = Recipe.findById(req.params.id)
+    const recipe = await Recipe.findById(req.params.id)
     if(!recipe) {
         res.status(400)
         throw new Error('Recipe not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // console.log(recipe)
+    // Make sure the logged in user matches the
+    // recipe user
+    if(recipe.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await recipe.deleteOne()
 
     res.status(200).json({id: req.params.id})
